Creates hands table
